텔레그램 봇 에러 수정
stop loss 값이 double 이라서 . 이 parsing 안되던 오류 수정
int 로 강제 형변환함
추가로 에러 body 로그 찍음

+++ src/main/java/minhobot/coincalculator/cci/CciScheduler.java
@@ -93,7 +93,7 @@ public class CciScheduler {
                     signal.equals("LONG") ? "🟢 LONG" : "🔴 SHORT",
                     price,
                     leverageResponse.getLeverage(),
-                    leverageResponse.getStoploss(),
+                    (int) leverageResponse.getStoploss(),
                     java.time.LocalDateTime.now()
             );
             telegramBotClient.sendMessage(text);
